If failure occurs when posting comment, don't close issue
https://github.com/probot/no-response/issues/24 describes a problem
where probot/no-response sometimes closes an issue without posting a
comment. We suspect that probot/no-response is attempting to post a
comment but is having the comment rejected by GitHub's abuse prevention
rate limits, similar to the issue experienced by probot/stale in
https://github.com/probot/stale/issues/26.

Following the approach implemented in
https://github.com/probot/stale/pull/64, this commit updates
probot/no-response to first attempt to post the comment saying that it
is closing the issue. If it successfully posts the comment, then it
closes the issue. If an error occurs when posting the comment, it does
not close the issue.